Refactor the datasets part, mainly.
Datasets part is now separated into modules, and 'next_batch' method
is implemented for the main program. Corresponding to this part, the
main 'models.py' program is slightly adjusted. What's more, an ex-
ternal program called 'aes.py' is created, which is to be the main
enter point of the whole program.

For model part, we are trying to apply Danijar's instruction to re-
factor it (see danijar.com/structuring-your-tensorflow-models),
which is just an experiment for now.